Convert 9 attestations to real collector-backed checks
The Collector framework lets controls move from "attestation_required"
to actually checking something. This batch converts 9 controls; manual
attestation count drops 79 → 70.

Auto via AdminClient broker config (no new collector data needed):

- KAFKA-AUTH-010 ssl.principal.mapping.rules set
- KAFKA-AUTH-013 delegation.token.secret.key set
- KAFKA-KRAFT-001 controller listener uses SASL or SSL
- KAFKA-KRAFT-002 controller listener uses TLS
- KAFKA-KRAFT-004 cluster.quorum_size >= 3
- KAFKA-AUDIT-001 authorizer.class.name set
- KAFKA-OPS-006 config.providers set

Auto via process collector (Linux /proc):

- KAFKA-OPS-007 open files ulimit >= 100k

Auto via tls collector (handshake + chain inspection):

- KAFKA-MON-004 leaf cert >= 30 days to expiry

Cluster.quorum_size cast to long so CEL int64 comparison works.

Sample run vs single-broker plaintext localhost:19092 with
adminclient+filesystem+tls collectors:
  Score: 0/100 | Pass: 24 | Fail: 24
    · 70 need operator attestation
    · 2 N/A — required collectors unavailable: process,jmx

+++ src/main/java/io/kafkascanner/collectors/KafkaCollectors.java
@@ -179,7 +179,7 @@ public final class KafkaCollectors {
             "controller_id", (long) controller.id(),
             "voters", nodes.stream().map(n -> (long) n.id()).toList(),
             "healthy", !nodes.isEmpty(),
-            "quorum_size", nodes.size(),
+            "quorum_size", (long) nodes.size(),
             "mode", mode
         );
     }
